test: split parser test into Slides and Sheets files (resolve max-lines warning)
After the previous de-sloppification commit (11ee4a0), the pre-commit
hook's Prettier step reformatted long object literals in
parserMatchingAndDocumentIds.test.js across multiple lines, bringing
the file from 496 to 528 lines and re-triggering the pre-existing
max-lines ESLint warning (at a lower line count). Resolves the
regression-checker's fingerprint-only regression on that warning.

Split the file structurally into two self-contained files:

- tests/parsers/slidesParserMatchingAndDocumentIds.test.js (313 lines,
  10 tests): the SlidesParser section verbatim with its own imports,
  setup/teardown, buildSlidesParserHarness, and slide/table/tagged
  element factories. Top-level describe: 'SlidesParser matching and
  document ID propagation'.
- tests/parsers/sheetsParserMatchingAndDocumentIds.test.js (219 lines,
  19 tests including one it.each with 3 cases): the SheetsParser
  section verbatim with its own setup/teardown, buildSheetsParserHarness,
  and SpreadsheetApp/TaskSheet mocks. Top-level describe: 'SheetsParser
  matching and document ID propagation'.
- tests/parsers/parserMatchingAndDocumentIds.test.js: deleted.

No shared helper extracted (KISS): the two sections shared no
genuinely shared logic — different refDocId constants, different mock
setup, different helper functions. Self-contained files chosen.

No test removed, renamed, merged, or parametrised. Every test name
and assertion from the original file is preserved exactly across the
two new files. Production code untouched. Code Reviewer: CLEAN.

Verification:
- Lint: 15 warnings (down from 16 baseline). The max-lines warning
  on the parser test is fully gone.
- Tests: 1870/1870 backend tests pass; 29 tests preserved (10 + 19).
- Pre-commit hook: Prettier clean, ESLint 0 errors, TypeScript build
  passes.
- Regression-checker: Regressions 0, New Failures 0, Fixes 1.
